Now get another access token if saved is invalid
I have include an option to be able to get another access token when the
saved one is invalid. It is necessary to add the constant
MAX_TRY_GET_VALID_ACCESS_TOKEN in Constants.java to be able to control
that. If you don't wan't to ask for another token you just need to fix
it to 0 otherwise you can set it a anything. For example I have set it
to 3 like this
public static final int MAX_TRY_GET_VALID_ACCESS_TOKEN = 3;
in Constants.java