Added some changes like key & improvement for token
Now it 's like before when MAX_TRY_GET_VALID_ACCESS_TOKEN = 0.
Added a small change in ecnryption process. Sure it's great to encrypt
but I think it's better if  key is different for each computer. Here we
have a chance on 2^127 to get the same key. And not every one is
changing of mac adress every day