Fix broken signature
ATLauncher was shipping with a JavaApplicationStub which also enforced
the contents of Info.plist. By modifying Info.plist, the signature was
therefore broken.

This led to it being a “broken signature” or “damaged application”,
which is something we should be training users to take seriously, not
just tell them to turn off signature checking entirely.

By removing the signature from this binary, you can now secondary-click
→ Open, and the app will ask you if it’s OK to open it, instead of
giving an error dialog.

Of course, it would be better for ATLauncher to be properly signed, but
I’m not its developer, so I can’t do anything about that.