misc: Rework Logging to use RxJava to lower memory usage.
Currently,
 the LauncherConsole keeps appending the log even when not visible.
This creates an issue when users don't even have the LauncherConsole
 visible, and it keeps chugging their memory.

This commit solves this issue by reworking logging.

First, LogManager is converted to use RXJava to pass along
 logs through the app.

Second, LoggingThread.java is deleted, this though leaves LogEvent.post
 orphaned and uncalled.

Third, Functionality of LogEvent.post is split between LauncherConsole
 and LogManager.

Fourth, LauncherConsole is adapted to use HierarchyFrame and to ignore
writes to the log when the UI is not visible.

The new issue presented by this change is that PasteUpload will not work
 as expected.
We can solve this instead by appending to a file then
 letting PasteUpload read from there.
 *This is not done in this commit.*

Helps #763